- Cleaned up

+++ app/src/main/java/com/zode64/trellodoing/widget/DoingWidget.java
@@ -18,32 +18,37 @@ import android.widget.Toast;
 
 import com.zode64.trellodoing.CardActionActivity;
 import com.zode64.trellodoing.CardAdderActivity;
-import com.zode64.trellodoing.utils.DoingNotification;
 import com.zode64.trellodoing.DoingPreferences;
 import com.zode64.trellodoing.KeepDoingActivity;
 import com.zode64.trellodoing.MainActivity;
 import com.zode64.trellodoing.R;
-import com.zode64.trellodoing.utils.TimeUtils;
-import com.zode64.trellodoing.utils.TrelloManager;
-import com.zode64.trellodoing.utils.WidgetAlarm;
 import com.zode64.trellodoing.db.ActionDAO;
+import com.zode64.trellodoing.db.AttachmentDAO;
 import com.zode64.trellodoing.db.BoardDAO;
 import com.zode64.trellodoing.db.CardDAO;
 import com.zode64.trellodoing.db.DeadlineDAO;
 import com.zode64.trellodoing.models.Action;
+import com.zode64.trellodoing.models.Attachment;
 import com.zode64.trellodoing.models.Board;
 import com.zode64.trellodoing.models.Card;
 import com.zode64.trellodoing.models.Member;
+import com.zode64.trellodoing.utils.DoingNotification;
+import com.zode64.trellodoing.utils.TimeUtils;
+import com.zode64.trellodoing.utils.TrelloManager;
+import com.zode64.trellodoing.utils.WidgetAlarm;
 
+import java.io.File;
 import java.util.ArrayList;
 import java.util.Calendar;
 import java.util.HashMap;
+import java.util.Iterator;
+import java.util.Map;
 
 import static com.zode64.trellodoing.utils.TimeUtils.between;
 
 public class DoingWidget extends AppWidgetProvider {
 
-    public final static String ACTION_REFRESH = "com.zode64.trellodoing.intent.action.REFRESH";
+    public final static String ACTION_SYNC = "com.zode64.trellodoing.intent.action.SYNC";
     public final static String ACTION_DEADLINE_ALARM = "com.zode64.trellodoing.intent.action.DEADLINE_ALARM";
     public final static String ACTION_STANDARD_ALARM = "com.zode64.trellodoing.intent.action.STANDARD_ALARM";
     public final static String ACTION_SET_ALARM = "com.zode64.trellodoing.intent.action.SET_ALARM";
@@ -51,21 +56,32 @@ public class DoingWidget extends AppWidgetProvider {
     public final static String ACTION_NETWORK_CHANGE = "com.zode64.trellodoing.intent.action.NETWORK_CHANGE";
     public final static String ACTION_ADD_CARD = "com.zode64.trellodoing.intent.action.ADD_CARD";
     public final static String ACTION_TODAY_BOARDS_SWITCH = "com.zode64.trellodoing.intent.action.TODAY_BOARDS_SWITCH";
+    public final static String ACTION_UPLOAD_ATTACHMENTS = "com.zode64.trellodoing.intent.action.UPLOAD_ATTACHMENTS";
 
     public static final String EXTRA_CARD_ID = "com.zode64.trellodoing.cardsproivder.EXTRA_CARD_ID";
     public static final String EXTRA_BOARD_ID = "com.zode64.trellodoing.cardsproivder.EXTRA_BOARD_ID";
+    public static final String EXTRA_WIFI_CHECKED = "com.zode64.trellodoing.cardsproivder.EXTRA_WIFI_CHECKED";
+
+    public static final int NO_CONNECTION = -1;
 
     private final static String TAG = DoingWidget.class.getName();
 
+    /**
+     * Prevents high frequency polling because of constant network changes
+     */
     private volatile static Boolean isAlreadyConnected = false;
+
+    private volatile static Integer connectionType = -2;
 
     @Override
     public void onUpdate( Context context, AppWidgetManager appWidgetManager, int[] appWidgetIds ) {
         Log.d( TAG, "onUpdate()" );
 
+        connectionType = connectionType( context );
+
         RemoteViews views = new RemoteViews( context.getPackageName(), R.layout.widget_doing );
         setSettingsListener( views, context );
-        setRefreshClickListener( views, context );
+        setSyncClickListener( views, context );
         setKeepDoingClickListener( views, context );
         setAddCardListener( views, context );
         setTodayBoardsSwitchListener( views, context );
@@ -101,10 +117,10 @@ public class DoingWidget extends AppWidgetProvider {
         widgetAlarm.stopDeadlineAlarm();
     }
 
-    private void setRefreshClickListener( RemoteViews views, Context context ) {
-        Intent refresh = new Intent( ACTION_REFRESH );
-        PendingIntent pendingRefresh = PendingIntent.getService( context, 0, refresh, 0 );
-        views.setOnClickPendingIntent( R.id.refresh, pendingRefresh );
+    private void setSyncClickListener( RemoteViews views, Context context ) {
+        Intent sync = new Intent( ACTION_SYNC );
+        PendingIntent pendingSync = PendingIntent.getService( context, 0, sync, 0 );
+        views.setOnClickPendingIntent( R.id.sync, pendingSync );
     }
 
     private void setKeepDoingClickListener( RemoteViews views, Context context ) {
@@ -146,6 +162,17 @@ public class DoingWidget extends AppWidgetProvider {
         views.setRemoteAdapter( resourceId, intent );
     }
 
+    public static int connectionType( Context context ) {
+        final ConnectivityManager connMgr = ( ConnectivityManager ) context
+                .getSystemService( Context.CONNECTIVITY_SERVICE );
+        NetworkInfo netInfo = connMgr.getActiveNetworkInfo();
+        if ( netInfo != null && netInfo.isConnected() ) {
+            return netInfo.getType();
+        } else {
+            return NO_CONNECTION;
+        }
+    }
+
     public static class NetworkChangeReceiver extends BroadcastReceiver {
 
         public NetworkChangeReceiver() {
@@ -155,17 +182,19 @@ public class DoingWidget extends AppWidgetProvider {
         @Override
         public void onReceive( final Context context, final Intent intent ) {
             Log.d( TAG, "onReceive()" );
-            final ConnectivityManager connMgr = ( ConnectivityManager ) context
-                    .getSystemService( Context.CONNECTIVITY_SERVICE );
-
-            NetworkInfo netInfo = connMgr.getActiveNetworkInfo();
             //should check null because in airplane mode it will be null
-            if ( netInfo != null && netInfo.isConnected() ) {
+            connectionType = connectionType( context );
+            if ( connectionType != NO_CONNECTION ) {
                 if ( !isAlreadyConnected ) {
                     // When the network returns any checks will begin as normal even if 'Keep Doing' alarm is set.
                     Log.d( TAG, "Connected to network" );
                     isAlreadyConnected = true;
                     context.startService( new Intent( ACTION_NETWORK_CHANGE, null, context, UpdateService.class ) );
+                }
+                if ( connectionType == ConnectivityManager.TYPE_WIFI ) {
+                    Intent attachmentsIntent = new Intent( ACTION_UPLOAD_ATTACHMENTS, null, context, UpdateService.class );
+                    intent.putExtra( EXTRA_WIFI_CHECKED, true );
+                    context.startService( attachmentsIntent );
                 }
             } else {
                 Log.d( TAG, "No network, stop checking for now" );
@@ -191,6 +220,10 @@ public class DoingWidget extends AppWidgetProvider {
 
             DoingPreferences preferences = new DoingPreferences( this );
             WidgetAlarm appWidgetAlarm = new WidgetAlarm( this.getApplicationContext(), preferences );
+            if ( ACTION_STOP_ALARM.equals( intent.getAction() ) ) {
+                appWidgetAlarm.stopStandardAlarm();
+                return;
+            }
             if ( ACTION_STOP_ALARM.equals( intent.getAction() ) ) {
                 appWidgetAlarm.stopStandardAlarm();
                 return;
@@ -225,38 +258,40 @@ public class DoingWidget extends AppWidgetProvider {
                 return;
             }
 
-            TrelloManager trelloManager = new TrelloManager( preferences.getSharedPreferences() );
+            TrelloManager trelloManager = new TrelloManager( preferences );
             CardDAO cardDAO = new CardDAO( this );
-
-            boolean isOnline = true;
-            // Start with out standing operations in the cache
             ActionDAO actionDAO = new ActionDAO( this, trelloManager, cardDAO );
-            ArrayList<Action> actions = actionDAO.all();
-            for ( Action action : actions ) {
-                if ( action.perform() ) {
-                    actionDAO.delete( action.getId() );
-                } else {
-                    isOnline = false;
-                    break;
-                }
-            }
+            ArrayList<Card> cards = null;
 
-            ArrayList<Card> cards = cardDAO.all();
-            if ( isOnline ) {
+            if ( connectionType != NO_CONNECTION ) {
+
+                // Start with out standing operations in the cache
+                ArrayList<Action> actions = actionDAO.all();
+                for ( Action action : actions ) {
+                    if ( action.perform() ) {
+                        actionDAO.delete( action.getId() );
+                    } else {
+                        break;
+                    }
+                }
+
+                if ( ACTION_UPLOAD_ATTACHMENTS.equals( intent.getAction() ) ) {
+                    if ( !intent.getBooleanExtra( EXTRA_WIFI_CHECKED, false ) ) {
+                        if ( connectionType == ConnectivityManager.TYPE_WIFI ) {
+                            return;
+                        }
+                    }
+                    postAttachments( trelloManager );
+                }
                 // Try update the cache
                 Member member = trelloManager.member();
                 if ( member != null ) {
-                    if ( ACTION_REFRESH.equals( intent.getAction() ) ) {
-                        BoardDAO boardDAO = new BoardDAO( this );
-                        boardDAO.deleteAll();
-                        for ( Board board : member.getBoards() ) {
-                            boardDAO.create( board );
-                        }
-                        boardDAO.closeDB();
+                    cards = member.getCards();
+                    if ( ACTION_SYNC.equals( intent.getAction() ) ) {
+                        sync( member );
                     }
                     // Clear cache
                     cardDAO.deleteAll();
-                    cards = member.getCards();
                     for ( Card card : cards ) {
                         cardDAO.create( card );
                         Log.d( TAG, "Card id: " + card.getId() );
@@ -264,21 +299,25 @@ public class DoingWidget extends AppWidgetProvider {
                 }
             }
 
+            cards = cardDAO.all();
             // Check to see if we should create any notifications
             DeadlineDAO deadlineDAO = new DeadlineDAO( this );
             HashMap<String, Long> deadlines = deadlineDAO.all();
             DoingNotification notifications = new DoingNotification( this );
             notifications.removeAll();
             int numDoingCards = 0;
-            String doingCardUrl = TrelloManager.TRELLO_URL;
+            String doingBoardUrl = preferences.getLastDoingBoard();
             Calendar now = Calendar.getInstance();
             for ( Card card : cards ) {
                 if ( card.isClockedOn() ) {
+                    doingBoardUrl = card.getBoardShortUrl();
+                    preferences.saveLastDoingBoard( doingBoardUrl );
                     numDoingCards++;
-                    doingCardUrl = card.getBoardShortUrl();
                     if ( !between( preferences.getStartHour(), preferences.getEndHour(), now ) ) {
-                        notifications.clockOff( doingCardUrl );
+                        notifications.clockOff( doingBoardUrl );
                     }
+                }
+                if ( card.isClockedOn() || card.getInListType() == Card.ListType.TODAY ) {
                     if ( TimeUtils.pastDeadline( deadlines.get( card.getServerId() ), now.getTimeInMillis() ) ) {
                         notifications.deadline( card.getBoardShortUrl() );
                     }
@@ -286,15 +325,16 @@ public class DoingWidget extends AppWidgetProvider {
             }
             if ( numDoingCards == 0 ) {
                 if ( between( preferences.getStartHour(), preferences.getEndHour(), now ) ) {
-                    notifications.clockOn( doingCardUrl );
+                    notifications.clockOn( doingBoardUrl );
                 }
             } else if ( numDoingCards > 1 ) {
-                notifications.multiDoings( doingCardUrl );
+                notifications.multiDoings( doingBoardUrl );
             }
 
             // Clean up
             actionDAO.closeDB();
             cardDAO.closeDB();
+            deadlineDAO.closeDB();
 
             AppWidgetManager manager = AppWidgetManager.getInstance( this );
             ComponentName thisWidget = new ComponentName( this, DoingWidget.class );
@@ -309,6 +349,68 @@ public class DoingWidget extends AppWidgetProvider {
             manager.notifyAppWidgetViewDataChanged( ids, R.id.doing_cards_list );
             manager.notifyAppWidgetViewDataChanged( ids, R.id.today_cards_list );
             manager.notifyAppWidgetViewDataChanged( ids, R.id.clocked_off_cards_list );
+        }
+
+        private void postAttachments( TrelloManager trelloManager ) {
+            AttachmentDAO attachmentDAO = new AttachmentDAO( this );
+            ArrayList<Attachment> pendingAttachments = attachmentDAO.allPending();
+            if ( !pendingAttachments.isEmpty() ) {
+                for ( Attachment attachment : pendingAttachments ) {
+                    int result = trelloManager.postAttachment( attachment );
+                    switch ( result ) {
+                        case TrelloManager.SUCCESS:
+                            attachmentDAO.setUploaded( attachment.getId() );
+                            break;
+                        case TrelloManager.FILE_NOT_FOUND:
+                            attachmentDAO.delete( attachment.getId() );
+                            break;
+                        default:
+                            // Do nothing
+                    }
+                }
+            }
+            attachmentDAO.closeDB();
+        }
+
+        private void sync( Member member ) {
+            Log.i( TAG, "Syncing..." );
+            ArrayList<Card> cards = member.getCards();
+            HashMap<String, Card> cardReg = new HashMap<>();
+            for ( Card card : cards ) {
+                cardReg.put( card.getServerId(), card );
+            }
+            AttachmentDAO attachmentDAO = new AttachmentDAO( this );
+            Iterator<Attachment> attachments = attachmentDAO.all().iterator();
+            while ( attachments.hasNext() ) {
+                Attachment attachment = attachments.next();
+                if ( !cardReg.containsKey( attachment.getCardServerId() ) ) {
+                    File file = attachment.getFile();
+                    File dir = file.getParentFile();
+                    file.delete();
+                    Log.i( TAG, "Deleting file:" + file.getName() );
+                    if ( dir.list().length == 0 ) {
+                        dir.delete();
+                        Log.i( TAG, "Deleting dir:" + dir.getName() );
+                    }
+                    attachmentDAO.delete( attachment.getId() );
+                }
+            }
+            attachmentDAO.closeDB();
+            DeadlineDAO deadlineDAO = new DeadlineDAO( this );
+            HashMap<String, Long> deadlines = deadlineDAO.all();
+            for ( Map.Entry<String, Long> deadline : deadlines.entrySet() ) {
+                if ( !cardReg.containsKey( deadline.getKey() ) ) {
+                    Log.i( TAG, "Deleting deadline for " + deadline.getKey() );
+                    deadlineDAO.delete( deadline.getKey() );
+                }
+            }
+            deadlineDAO.closeDB();
+            BoardDAO boardDAO = new BoardDAO( this );
+            boardDAO.deleteAll();
+            for ( Board board : member.getBoards() ) {
+                boardDAO.create( board );
+            }
+            boardDAO.closeDB();
         }
 
         public class DisplayTokenMissingToast implements Runnable {
